Fix bug on navigation dots
When the first navigation dot is clicked and then the user clicks next, it returned an error. The issue was that the index taken from the class name is a string and when moveForward function is run and tries to add 1 to "0", it caused an error. This was fixed by casting the index to an integer.