Slice starting from 7th index
Previously, it took the index from the last character of the class name. This will limit the carousel to have less than 10 images. To fix this, I made it start slicing from the 7th index since for all images the first seven characters are the same, it's going to always take the number after hyphen symbol.